Models, Controllers, Views sorted, CMakeLists.txt changed
