+++ client/CMakeLists.txt
@@ -6,7 +6,7 @@ set(CMAKE_INCLUDE_CURRENT_DIR ON)
 
 if (NOT QT_MAJOR_VERSION)
     set(QT_MAJOR_VERSION 6)
-endif()
+endif ()
 
 find_package(Qt${QT_MAJOR_VERSION}Widgets REQUIRED)
 find_package(Qt${QT_MAJOR_VERSION}Quick REQUIRED)
@@ -15,17 +15,20 @@ set(CMAKE_AUTOMOC ON)
 set(CMAKE_AUTOUIC ON)
 set(CMAKE_AUTORCC ON)
 
-include_directories(${CMAKE_BINARY_DIR})
+include_directories(${CMAKE_BINARY_DIR}
+        controllers
+        models
+        views)
 
 # https://api.kde.org/frameworks-api/frameworks-apidocs/frameworks/karchive/html/index.html
 find_package(KF5Archive REQUIRED)
 
 add_executable(client
         main.cpp
-        cavokeclientcontroller.cpp
-        cavokeclientmodel.cpp
-        cavokeclientview.cpp
-        cavokeqmlgamemodel.cpp
+        controllers/cavokeclientcontroller.cpp
+        models/cavokeclientmodel.cpp
+        views/cavokeclientview.cpp
+        models/cavokeqmlgamemodel.cpp
         tictactoelogic.cpp
         cache_manager.cpp)
